--- src/engine/index.ts
@@ -1,8 +1,0 @@
-// Core session functionality
-export * from '../core';
-export * from '../workers';
-
-// Agent workflow components
-export * from '../workflow';
-export * from '../processing';
-
