--- src/processing/prompts/index.ts
@@ -1,3 +1,0 @@
-// // Prompt building and templates
-// export { PromptBuilder } from './builder';
-// export { BasePromptTemplate, PROMPT_TEMPLATES, type PromptTemplate } from './templates';
