+++ src/workflow/index.ts
@@ -2,11 +2,8 @@
 export { WorkflowExecutor } from './executor';
 export { StepExecutor } from './step-executor';
 
-// Memory management
-export { MemoryOptimizer } from './memory/memory-optimizer';
-
 // State management
-export { WorkflowStateManager, type WorkflowExecutionState } from './workflow-state';
+export { WorkflowStateManager } from './workflow-state';
 
 // Result builders
 export { WorkflowResultBuilder } from './result-builder';
